refactor: 대시보드 패키지 구조 변경
