+++ src/main/java/com/codeit/team4/deokhugam/dashboard/entity/PopularReview.java
@@ -1,6 +1,4 @@
-package com.codeit.team4.deokhugam.dashboard.review;
-
-import com.codeit.team4.deokhugam.dashboard.entity.PeriodType;
+package com.codeit.team4.deokhugam.dashboard.entity;
 
 import com.codeit.team4.deokhugam.book.entity.Book;
 import com.codeit.team4.deokhugam.global.entity.BaseEntity;
