+++ src/main/java/com/codeit/team4/deokhugam/dashboard/repository/PopularReviewRepository.java
@@ -1,4 +1,6 @@
-package com.codeit.team4.deokhugam.dashboard.review;
+package com.codeit.team4.deokhugam.dashboard.repository;
+
+import com.codeit.team4.deokhugam.dashboard.entity.PopularReview;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
