+++ src/main/java/com/codeit/team4/deokhugam/dashboard/repository/PowerUserRepository.java
@@ -1,4 +1,6 @@
-package com.codeit.team4.deokhugam.dashboard.user;
+package com.codeit.team4.deokhugam.dashboard.repository;
+
+import com.codeit.team4.deokhugam.dashboard.entity.PowerUser;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
