+++ src/test/java/com/codeit/team4/deokhugam/dashboard/entity/PopularBookTest.java
@@ -1,6 +1,4 @@
-package com.codeit.team4.deokhugam.dashboard.book;
-
-import com.codeit.team4.deokhugam.dashboard.entity.PeriodType;
+package com.codeit.team4.deokhugam.dashboard.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
