+++ src/test/java/com/codeit/team4/deokhugam/dashboard/service/PopularBookQueryBuilderTest.java
@@ -1,4 +1,4 @@
-package com.codeit.team4.deokhugam.dashboard.book;
+package com.codeit.team4.deokhugam.dashboard.service;
 
 import static com.codeit.team4.deokhugam.jooq.tables.Books.BOOKS;
 import static com.codeit.team4.deokhugam.jooq.tables.Reviews.REVIEWS;
